made the changes in login

+++ frontend/src/app/core/services/auth.service.ts
@@ -1,49 +1,61 @@
-import { Injectable } from '@angular/core';
-import {HttpClient} from '@angular/common/http';
-import {Router} from '@angular/router';
+import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { isPlatformBrowser } from '@angular/common';
+import { Router } from '@angular/router';
 import { LoginResponse, RegisterResponse, User } from '../models/auth.model';
-import { Observable,tap } from 'rxjs';
+import { Observable, tap } from 'rxjs';
 
 @Injectable({
   providedIn: 'root',
 })
 export class AuthService {
-  private baseUrl='http://localhost:3000/api/auth';
+  private baseUrl = 'http://localhost:3000/auth';
 
   constructor(
-    private http:HttpClient,
-    private router:Router
-  ){}
+    private http: HttpClient,
+    private router: Router,
+    @Inject(PLATFORM_ID) private platformId: Object
+  ) {}
 
-  login(data:{email:string;password:string}):Observable<LoginResponse>{
-    return this.http.post<LoginResponse>(`${this.baseUrl}/login`,data)
-    .pipe(
-      tap((res)=>{
-        localStorage.setItem('authToken',res.token);
-        localStorage.setItem('currentUser',JSON.stringify(res.user));
-      })
-    );
-}
+  login(data: { email: string; password: string }): Observable<LoginResponse> {
+    return this.http.post<LoginResponse>(`${this.baseUrl}/login`, data)
+      .pipe(
+        tap((res) => {
+          if (isPlatformBrowser(this.platformId)) {
+            localStorage.setItem('authToken', res.token);
+            localStorage.setItem('currentUser', JSON.stringify(res.user));
+          }
+        })
+      );
+  }
 
-register(data:{name:string;email:string;password:string}):Observable<RegisterResponse>{
-  return this.http.post<RegisterResponse>(`${this.baseUrl}/register`,data);
-}
+  register(data: { name: string; email: string; password: string }): Observable<RegisterResponse> {
+    return this.http.post<RegisterResponse>(`${this.baseUrl}/register`, data);
+  }
 
-logout(){
-  localStorage.clear();
-  this.router.navigate(['/login']);
-}
+  logout() {
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.clear();
+    }
+    this.router.navigate(['/login']);
+  }
 
-getToken():string| null{
-  return localStorage.getItem('authToken');
-}
+  getToken(): string | null {
+    if (isPlatformBrowser(this.platformId)) {
+      return localStorage.getItem('authToken');
+    }
+    return null;
+  }
 
-getUser(): User | null{
-  const user=localStorage.getItem('currentUser');
-  return user?JSON.parse(user):null;
-}
+  getUser(): User | null {
+    if (isPlatformBrowser(this.platformId)) {
+      const user = localStorage.getItem('currentUser');
+      return user ? JSON.parse(user) : null;
+    }
+    return null;
+  }
 
-isLoggedIn():boolean{
-  return !!this.getToken();
-}
+  isLoggedIn(): boolean {
+    return !!this.getToken();
+  }
 }
